requirements.txt
openai
streamlit
pandas
newspaper3k